16- Correção rotas.

+++ src/Controller/VeiculosController.php
@@ -169,14 +169,21 @@ class VeiculosController extends AppController
      * @return \Cake\Http\Response|null Redirects to index.
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function delete($id = null)
+    public function delete()
     {
         $this->request->allowMethod(['post', 'delete']);
 
         $response = null;
         $statusCode = 200;
 
+        // Pegando o id do corpo da requisição (POST)
+        $id = $this->request->getData('id');
+
         try {
+            if (!$id) {
+                throw new \Exception('ID não fornecido.');
+            }
+
             $veiculo = $this->Veiculos->get($id);
 
             if ($this->Veiculos->delete($veiculo)) {
@@ -208,5 +215,6 @@ class VeiculosController extends AppController
             ->withStatus($statusCode)
             ->withStringBody(json_encode($response));
     }
+
 
 }
